🩹 [Patch]: Remove other modules before reimporting reqs for processed module (#80)
## Description


This pull request includes changes to the `scripts/helpers/Build`
directory, focusing on improving the handling of module outputs and
dependencies. The most important changes are as follows:

Improvements to module output handling:

*
[`scripts/helpers/Build/Build-PSModuleRootModule.ps1`](diffhunk://#diff-1d337ff39f37506a54fda1c5d0487f1b2c2ef318f216a4d9a56c3e7248b69879R39-R42):
Suppressed the 'PSAvoidUsingWriteHost' warning to allow direct console
output instead of writing to the pipeline.
*
[`scripts/helpers/Build/Build-PSModuleRootModule.ps1`](diffhunk://#diff-1d337ff39f37506a54fda1c5d0487f1b2c2ef318f216a4d9a56c3e7248b69879L245-R249):
Changed `Show-FileContent` calls to use `Write-Host` for direct console
output in the 'Build root module - Result' sections.
[[1]](diffhunk://#diff-1d337ff39f37506a54fda1c5d0487f1b2c2ef318f216a4d9a56c3e7248b69879L245-R249)
[[2]](diffhunk://#diff-1d337ff39f37506a54fda1c5d0487f1b2c2ef318f216a4d9a56c3e7248b69879L256-R260)

Enhancements to module dependency handling:

*
[`scripts/helpers/Build/Import-PSModule.ps1`](diffhunk://#diff-15f2301e39f56f54fa469db56cc7c487e2df36eb295bb74e8d2325aabcd164c0L31-R34):
Added logic to remove required and nested modules of an existing module
before re-importing it, ensuring a clean state.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas