🌟 [Major]: Moved docs generation to `Document-PSModule` (#105)
## Description

This pull request removes the functionality related to building and
uploading module documentation, by removing the
`Build-PSModuleDocumentation` function and the associated artifact
upload step in the GitHub Actions workflow.

This functionality is moved to
[`Document-PSModule`](https://github.com/PSModule/Document-PSModule/)

Removal of documentation build and upload:

*
[`scripts/helpers/Build/Build-PSModuleDocumentation.ps1`](diffhunk://#diff-dc95bfbd255c5bb79df192179e9cc5141ef230a8d82e80bc12fc6fefe6186fffL1-L110):
Entire file deleted, removing the `Build-PSModuleDocumentation` function
which was responsible for generating and fixing markdown documentation
for the module.
*
[`scripts/helpers/Build-PSModule.ps1`](diffhunk://#diff-c688e346ad60fbe881bd05b6a5dbc1cd712fdca0bf8b8527965db583af825fd9L55):
Removed the call to `Build-PSModuleDocumentation` from the
`Build-PSModule` function.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L74-L81):
Removed the step that uploads the documentation artifact in the GitHub
Actions workflow.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas