🪲 [Fix]: Fix `Variables` being `$null` (#110)
## Description

This pull request includes a minor change to the
`Get-PSModuleVariablesToExport.ps1` script. The change involves moving
the initialization of the `$variablesToExport` variable to an earlier
point in the script to ensure it is defined before any potential early
returns.

*
[`scripts/helpers/Build/Get-PSModuleVariablesToExport.ps1`](diffhunk://#diff-e2148fcf63283d70b3d432f67fc98191e2cc712dc3f4d30914c8c0810d11a6deR28-L35):
Moved the initialization of `$variablesToExport` to before the folder
existence check to ensure it is always defined.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas